doctor-service/requirements.txt
Django>=4.2
djangorestframework
django-environ
requests
psycopg2-binary
djangorestframework-simplejwt
